Support AsciiDoc manual files
These files will be converted into DocBook 5.0 before the manual is
built in the standard manner.

The AsciiDoc files should be within the `documentsDirectory` and have
the file suffix `adoc`. The resulting DocBook file will have the same
path within the combined documents directory but with the file suffix
`xml`.