Convert directly to DocBook 5
Unfortunately the asciidoc tool does not define the `xlink` namespace
when omitting the header output. We therefore inject the definition
manually. This is done using sed, which may not be the best idea in
the world but since we have more or less full control over the XML it
should be reasonably robust.