Update HTML documentation style
This is a complete overhaul of the HTML documentation styling. The new
style is custom made from scratch and includes support for
`prefers-color-scheme`.

To develop the CSS file in a reasonable way we use SCSS through the
light-weight rsass tool (https://github.com/kaj/rsass). We also add
the Flake development commands `p-build` to build the final CSS file
and `p-watch` to continuously build it when the source files change.
Note, we include the final CSS file in the repository to avoid
introducing rsass as a dependency for end users.

We also update highlight.js to the latest version and include a plugin
to retain DocBook callouts.