Fix options-db.xml on darwin
See https://github.com/NixOS/nix/pull/4276

Without this, Nix tries to stat channels in /nix/var/nix and fails.
This happens when determining the default value of the `nix-path`
option, so before we have a chance to override `nix-path` or set
`restrict-eval`; our only hope is to confine the looked-up paths to the
sandbox.