Only pass --store dummy:// if it is supported by Nix
Nix < 2.4 does not support the dummy store, so we'll need to hope
setting NIX_STORE_DIR to a temporary one is enough. This allows using
home-manager's documentation facilities with Nix 2.3 again.

To verify I've built the home-manager's docs attribute both normally and
with an overlay that switches out `pkgs.nix` for `pkgs.nix_2_3` (which is
how we've run into this issue originally).